requirements.txt
alembic==1.13.1
black==23.11.0
# Core Framework
fastapi==0.104.1
flake8==6.1.0

# AI & NLP - Free Tier Focus
google-generativeai==0.3.2  # Google Gemini - generous free tier

# HTTP Client
httpx==0.25.2
isort==5.13.2

# Logging & Monitoring
loguru==0.7.2
numpy==1.25.2
openai==1.3.6  # OpenAI - for backup

# Data Processing
pandas==2.1.4
passlib[bcrypt]==1.7.4

# PDF Processing
pdfplumber==0.10.0
pre-commit==3.6.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
sentence-transformers==2.2.2  # For embeddings
spacy==3.7.2  # Advanced NLP

# Database & ORM
sqlalchemy==2.0.23
torch==2.1.1  # PyTorch for transformers
transformers==4.36.0  # Hugging Face transformers
uvicorn[standard]==0.24.0
